loop system init and test
