+++ src/index.js
@@ -76,7 +76,6 @@ const moment = require("moment");
 var Long = require("long");
 const chalk = require("chalk");
 const Path = require('path')
-const schedule = require('node-schedule');
 
 // Internal modules
 const automod = require("./automod");
@@ -93,6 +92,7 @@ const crypto = require("./crypto");
 const profile = require("./profile");
 const wolfram = require("./wolfram");
 const modlog = require("./events/modlog");
+const loops = require("./mainloops");
 
 // URL that points to MongoDB database
 var url = "mongodb://localhost:27017/zora";
@@ -212,53 +212,6 @@ const ServerM = mongoose.model("Servers", defaultConfig);
 // var memeInterval = setInterval(getMemes, config.reddit.interval * 1000 * 60 * 60);
 
 client.on("ready", async function () {
-
-  /**
-   * Start the main loop, this runs every hour
-   * used to update things such as stats
-   */
-  /*
-  async function update() {
-    // Get from database and sort!
-    const getSort = () => {
-      return UserM.find({}).sort({
-        zcoins: -1
-      }).exec()
-    }
-
-    var sorted = await getSort();
-
-    // Default to 100
-    var top = 25;
-
-    // Add fields
-    var counter = 1;
-    for (var usr in sorted) {
-      var profile = sorted[counter - 1];
-      if (profile) {
-        if (counter <= top && counter <= 25) {
-          if (counter === 1) {
-            client.guilds.forEach(async function (guild) {
-              await ServerM.findById(guild.id, function (err, server) {
-                console.log('update_data: ' + guild.id)
-                server.stats.richest.id = profile._id;
-                server.stats.richest.name = profile.username;
-                server.stats.richest.zcoins = profile.zcoins;
-                server.stats.richest.level = profile.level;
-                server.save();
-              });
-            });
-          }
-        }
-      }
-    }
-  }
-
-  update();
-  var j = schedule.scheduleJob('0 * * * *', async function () {
-    update();
-  });
-*/
 
   // Set activity
   setInterval(function () {
@@ -684,6 +637,9 @@ client.on("message", async message => {
       UserM.findById(message.author.id, function (err, user) {
 
         admin.bot(client, message, command, args, defaultConfig, defaultprofile, user, cserver, UserM, ServerM);
+
+        // LOOOOPSSSS BOI!!!
+        loops.bot(client, message, command, args, user, cserver, UserM, ServerM);
 
         // Weather
